+++ R/heading.R
@@ -3,9 +3,9 @@
 #'
 #' Determines the heading between spatial points.
 #'
-#' @param from The starting position; an object of class SpatVector
+#' @param from The starting position; an object of class `SpatVector`.
 #'
-#' @param to The ending position;  an object of class SpatVector
+#' @param to The ending position; an object of class `SpatVector`.
 #'
 #' @return The heading between the points, in degrees.
 #'
@@ -57,7 +57,16 @@ heading <- function(from, to) {
 }
 
 coords <- function(crds) {
-  if (inherits(crds, "SpatVector")) {
+  if (is.matrix(crds)) {
+    if (isS4(crds)) {
+      crds <- crds@.Data[, 1:2, drop = FALSE]
+      if (!identical(colnames(crds), xycolNames))
+        colnames(crds) <- xycolNames
+    } else {
+      crds <- crds[, 1:2, drop = FALSE]
+    }
+
+  } else if (inherits(crds, "SpatVector")) {
     .requireNamespace("terra")
     crds <- terra::crds(crds)
   } else if (inherits(crds, "sf")) {
@@ -72,14 +81,22 @@ coords <- function(crds) {
 }
 
 `coords<-` <- function(obj, value) {
-  if (inherits(obj, "SpatVector")) {
+  if (is.matrix(obj)) {
+    if (isS4(obj)) {
+      obj@.Data[, 1:2] <- value
+    } else {
+      obj[, 1:2] <- value
+    }
+  } else if (inherits(obj, "SpatVector")) {
     .requireNamespace("terra")
-    crdsdf <- data.frame(value, as.data.frame(coords(obj)))
-    colnames(crdsdf) <- c("x", "y", "x1", "y1")
-    obj <- terra::vect(crdsdf, geom = c("x", "y"))
+    crdsdf <- data.frame(value, as.data.frame(obj))
+    if (!identical(colnames(crdsdf)[1:2], xycolNames)) {
+      colnames(crdsdf)[1:2] <- xycolNames
+    }
+    obj <- terra::vect(crdsdf, geom = xycolNames)
   } else if (inherits(obj, "sf")) {
     .requireNamespace("sf")
-    obj2 <- sf::st_as_sfc(sf::st_as_sf(as.data.frame(value), coords = c("x", "y")))
+    obj2 <- sf::st_as_sfc(sf::st_as_sf(as.data.frame(value), coords = xycolNames))
     obj <- sf::st_set_geometry(obj, value = obj2)
     obj
   } else if (inherits(obj, "Spatial")) {
@@ -89,3 +106,6 @@ coords <- function(crds) {
 
   obj
 }
+
+x1y1colNames <- c("x1", "y1")
+xycolNames <- c("x", "y")