+++ R/spades-tools-package.R
@@ -20,7 +20,7 @@
 #'   [adj()] \tab An optimized (i.e., faster) version of
 #'                          [terra::adjacent()]\cr
 #'   [cir()] \tab Identify pixels in a circle around a
-#'                          [`SpatialPoints*()`][sp::SpatialPoints-class] object\cr
+#'                          `SpatialPoints*` object\cr
 #'   [directionFromEachPoint()] \tab Fast calculation of direction and
 #'                                             distance surfaces\cr
 #'   [distanceFromEachPoint()] \tab Fast calculation of distance surfaces\cr
@@ -37,7 +37,7 @@
 #' \tabular{ll}{
 #'   [crw()] \tab Simple correlated random walk function\cr
 #'   [heading()] \tab Determines the heading between `SpatialPoints*`\cr
-#'   quickPlot::makeLines() \tab Makes `SpatialLines` object for, e.g., drawing arrows\cr
+#'   `quickPlot::makeLines()` \tab Makes `SpatialLines` object for, e.g., drawing arrows\cr
 #'   [move()] \tab A meta function that can currently only take "crw"\cr
 #'   [specificNumPerPatch()] \tab Initiate a specific number of agents per patch\cr
 #' }
@@ -48,7 +48,7 @@
 #' contributed packages such as `sp`, `raster`, `maps`, `maptools`
 #' and many others), we provide the following GIS-related functions:
 #' \tabular{ll}{
-#'   quickPlot::equalExtent() \tab Assess whether a list of extents are all equal\cr
+#'   `quickPlot::equalExtent()` \tab Assess whether a list of extents are all equal\cr
 #' }
 #'
 #' @section 4 Map-reduce - type operations: