+++ tests/testthat/test-adj.R
@@ -1,7 +1,6 @@
 test_that("adj.R results not identical to adjacent", {
-  # skip_if_not_installed(c("raster", "terra"))
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
   data.table::setDTthreads(1)
 
   for (ii in seq(NROW(rastDF))) {
@@ -121,7 +120,7 @@ test_that("adj.R results not identical to adjacent", {
 
 test_that("errors in adj are not correct", {
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
   data.table::setDTthreads(1)
 
   for (ii in seq(NROW(rastDF))) {
@@ -140,7 +139,7 @@ test_that("errors in adj are not correct", {
 test_that("adj.R: torus does not work as expected", {
 
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
   data.table::setDTthreads(1)
 
   for (ii in seq(NROW(rastDF))) {